+++ lib/quantization.cpp
@@ -23,7 +23,7 @@ constexpr float qscale[64] = {
 // clang-format on
 
 void create_qtable(int c, int QF, int *qtable) {
-  float scale = (QF < 50) ? 5000.0F / QF : 200.0F - 2.0F * QF;
+  float scale = (QF < 50) ? 5000.0F / static_cast<float>(QF) : 200.0F - 2.0F * static_cast<float>(QF);
   for (int i = 0; i < 64; ++i) {
     float stepsize = (qmatrix[c][i] * scale + 50.0F) / 100.0F;
     int val;
